# specs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "AI4I-Core",
      "description": "A comprehensive microservices based platform to handle language based AI models inferences at scale.",
      "version": "1.0.0"
    },
    "paths": {
...
      "/api/v1/auth/validate": {
        "get": {
          "tags": [
            "Authentication"
          ],
          "summary": "Validate Token",
          "description": "Validate token",
          "operationId": "validate_token_api_v1_auth_validate_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/XAuthSource"
            }
          ]
        }
      },
...
      "/api/v1/auth/permissions": {
        "get": {
          "tags": [
            "Authentication",
            "Admin"
          ],
          "summary": "Get All Permissions",
          "description": "Get all permissions from the permissions table (Admin only)\n\nReturns a list of all permissions with:\n- id\n- name\n- resource\n- action\n- created_at",
          "operationId": "get_all_permissions_api_v1_auth_permissions_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/XAuthSource"
            }
          ]
        }
      },
...
    },
    "components": {
...
      "securitySchemes": {
        "HTTPBearer": {
          "type": "http",
          "scheme": "bearer"
        },
        "APIKeyHeader": {
          "type": "apiKey",
          "in": "header",
          "name": "X-API-Key"
        }
      },
      "parameters": {
        "XAuthSource": {
          "name": "x-auth-source",
          "in": "header",
          "required": false,
          "description": "Select auth source for testing in Swagger (API_KEY or AUTH_TOKEN)",
          "schema": {
            "type": "string",
            "enum": [
              "API_KEY",
              "AUTH_TOKEN"
            ],
            "default": "API_KEY"
          }
        }
      }
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "API Gateway (with API Key Permission Validation)"
      }
    ],
...
  }